# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/auth/login": {
      "post": {
        "operationId": "AdminAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "maxLength": 1024
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated session"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Too many attempts"
          }
        },
        "summary": "Create an administrator browser session",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/v1/admin/auth/session": {
      "get": {
        "operationId": "AdminAuthController_session",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "rn_admin_session": []
          }
        ],
        "summary": "Read the current administrator session",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/v1/admin/auth/logout": {
      "post": {
        "operationId": "AdminAuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "rn_admin_session": []
          }
        ],
        "summary": "Revoke the current administrator session",
        "tags": [
          "admin-auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "RN Foundation API",
    "description": "Mobile bootstrap, localization, theme and update policy",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rn_admin_session"
      }
    },
...
  }
}